Create package.json
package.json is a configuration file for a Node.js project. It specifies the package dependencies for the project, as well as scripts that can be run to perform various tasks.

The package.json file is typically located in the root directory of a Node.js project and is used by the npm package manager to manage the project's dependencies.