Create deploy.js
deploy.js is a custom script that can be used with the Hardhat Ethereum development framework to deploy contracts or perform other tasks related to contract deployment.

Hardhat is a popular tool for developing and testing Ethereum smart contracts and decentralized applications (DApps). It provides a number of features and plugins that make it easy to work with Ethereum, including a built-in development network, support for various Solidity compilers, and the ability to define custom scripts.

To use a deploy.js script with Hardhat, you will need to include it in your hardhat.config.js file under the scripts field. The script should be a JavaScript function that takes a HardhatRuntimeEnvironment object as an argument. This object provides access to various utility functions and objects that can be used in the script, such as the deploy function for deploying contracts.